requirements.txt
yt-dlp==2025.9.26
pydantic==2.12.2
pydantic-settings==2.11.0
scikit-learn==1.3.2
faiss-cpu==1.8.0.post1
filelock==3.15.4
langchain-ollama==0.1.3
pytest==8.3.3
numpy>=1.24,<2.0

